Save 8 bit b/w images using PseudoGrey Plus
PseudoGrey Plus by Rich Franzen is an algorithm to store 3110 levels of
gray in 24 bit color images. This is simply done by adding 1 or 2 to the
R, G, and/or B values to adjust the pixel's luminance. It's a subtle but
perceivable effect and helps to prevent banding.

This implementation adds a new method isBW() to IImage, which is only
implemented for Image16 and defaults to false. It's only called once per
saving and has a shortcut to determine that the image is NOT b/w.

The minimally invasive core algorithm was added to rtengine/imageio.cc
and is only called when saving an 8 bit image that is b/w.

Please note that when saving as JPEG, compression squashes the effect
and a quality of 95+ and 4:4:4 subsampling is recommended. Many
PseudoGrey sample image on the net suffer from this and do not really
show the benefit.