See #3446: the theme filenames now includes a code for version handling.
The theme filename must use a template, that will be explained with this
example:

mytheme-GTK3-20_22.css

Everything before -GTK3... is the theme (short) name, the [-GTK3-] is a
mandatory part, 20 is the included version where support begins and 22
is the included version where the support ends. One of this two version
parameters can be ommited (but not both of course), which make RT
understand "is compatible from all version up to xx" (e.g.
mytheme-GTK3-_22.css) or "is compatible from version yy and above" (e.g.
mytheme-GTK3-20_.css)

The RawTherapee-GTK3-_19.css comes 'as is' from the gtk3 branch and has
still to be tuned.