simplify the implementation of MyFileChooserButton::show_chooser
Do not try to show a native file dialog. I think this would be useless, since
RT already uses non-native dialogs in other places, and I'm not sure it would
be possible to replace all of them (e.g. in icmpanel.cc the dialog is
customized by adding a checkbox for WB, and this is not possible with a native
dialog).