Move r pix coords to arguments of amaze
This is an effort to modularize some of the RawTherapee functions so
they can be reused in a library to share with other open-source photo
editors.

The first step is to make explicit the dependencies needed by
amaze_demosaic_RT.

It compiles and seems to work normally, but I haven't tested it on pixel
shift stuff.