Review `FileCatalog::_refreshProgressBar()`
This is an experimental change for the `FileCatalog` which introduces
a `#define` named `USE_IR` which switches between the old, sequential
behavior when loading the catalog and a new faster one, which uses
the `IdleRegister` for `FileCatalog::previewReady()` (see #5312).

I also experimented with a `FileCatalog::_refreshProgressBar()` relying
completely on `IdleRegister` (and tracking already pending calls) but
as there isn't enough idle time it only showed "100%" without any
progress.

The problem is that there's either a visible progress which slows the
whole processing down or it's blazingly fast and there's no visible
progress and an unresponsiveness while all the thumbs are rendered at
once.