CI: Fix CVE-2020-15228 in GitHub Actions (#6004)
* CI: Fix CVE-2020-15228 in GitHub Actions

Fixes the CVE-2020-15228 security issue in GitHub Actions:
> `add-path` and `set-env` Runner commands are processed via stdout
set-env was used twice, which is now replaced by  >> $GITHUB_ENV

 - Issue: https://github.com/actions/toolkit/security/advisories/GHSA-mfwh-5m23-j46w
 - Solution: https://docs.github.com/en/free-pro-team@latest/actions/reference/workflow-commands-for-github-actions#setting-an-environment-variable)

* CI: Specify installing libsigc++ v2.x in GitHub Actions

Version 3 of libsigc++ was automatically installed. This commit fixates the version brew installs to 2.x.

Should fix #6002

* CI: Use v2 checkout and upload-artifact actions

Also replaces macos-10.15 with macos-latest as runner. Currently is references the same environment (macOS 10.15), but this way it will update to macOS 11.0 when it's ready and stable.