Add ORI to the recognized file formats
Disabled by default as .ori extension might be
used to identify a file as "original".

When shooting Olympus High Res files the camera
saves the High Res file as .orf and the first
of the 8 frames as .ori.
The .ori can later be used to fix artefacts from
motion or as a backup if the whole High Res came
out flawed.