Create a preferences widget for external editors
The widget is intended to be embedded in the preferences dialog. It
displays a list of external editors using an icon, name, and command for
each editor. The name and command can be edited. Editors can be added,
removed, and rearranged. A button allows one to set an editor's
information by selecting from a list of installed applications.