Fix tmp image reuse when exporting different image
If two images have identical processing parameters, then sending one to
an external editor followed by sending the other one will cause the
first temporary image to be reused. This commit associates the image in
the editor with the "cached" temporary image so that the temporary image
only gets used if the editor image matches.