Allow really tiny spots for removal (#6418)
I may be a very picky person, but I want to repair really tiny specks of dust on my negative scans.
I tried setting the SpotParams::minRadius to 1 and this seems to work fine. I can zoom in in the
UI and still manipulate things with point and click. I guess 1 is a natural lower boundary for the
radius that should be explored;-)