Default strings cleanup, includes:
1 - Obsolete history events removal

Remove history events that are not referenced anywhere in code

Left event 149 despite being unreferenced due to the comment in the code indicating it may return

2 - Orphaned strings removal

HISTOGRAM_TOOLTIP_RAW appears to be an orphan, but this looks strange to me.  Investigation needed

3 - Strings cleanup - remove improperly commented entries

These don't get treated as comments, but at least didn't break anything since
nothing in the code referenced them.

But they will waste translator's time, so delete them