Move tone equalizer after exposure
Place it between the shadows curve (black and shadow compression) and
DCP tone curve and look table. Moving it after exposure makes it
possible to easily adjust the exposure to the ball-park range before
applying the tone equalizer. Keeping it before the tone curves allows it
to affect the image before clipping occurs.