Add basic camconst entry for Canon EOS R8
The camera seems to have the same characteristics as the Canon EOS R6m2.
The raw crops for the R6m2 are derived from standard and CRAW images for
full-frame and 1.6 crop. The raw crops for the R8 are from standard and
CRAW images for full-frame only, but for both standard and Dual Pixel.
The Dual Pixel R8 images require 10 pixels cropped from the right side
instead of 6. Thus, the full-frame crop needs adjustment.