Display actual language names in preferences
Add ability to set metadata variables in the language files.
Use the LANGUAGE_DISPLAY_NAME metadata variable value in preferences to
show the languages.

Metadata variables are defined in the first section of a language file,
i.e. the comments at the beginning of the file. At most one variable can
be defined per line. The definition takes the following format:
    #<header><whitespace>@<key>=<value>
where:
    <header> is any sequence of non-whitespace characters
    <whitespace> is any non-empty sequence of whitespace characters
    <key> is the metadata key name
    <value> is the metadata value
For example,
    #101 @LANGUAGE_DISPLAY_NAME=English (US)
defines a metadata key "LANGUAGE_DISPLAY_NAME" with value
"English (US)". Lines that do not match the format are ignored.