Fix auto DCP detection when using US English
For systems where the locale is not English but the language in
preferences is manually set to "English (US)", the Color Management
Auto-matched camera profile option is not selectable even if a DCP
exists for the camera. This is because the Glib::ustring
casefold_collate_key() method returns a locale-dependent value. When
initializing the list of DCP files, the system locale is used. When
finding a matching DCP profile for an image, a different locale is used
because the LANG environment variable is set while initializing the
RawTherapee GUI language.