cli: don't silently ignore -- options
in the GUI we call gtk_init with argc/argv, and therefore silently ignore
any --long-form CLI options on the assumption that GTK will deal with them.

In the CLI we do not call gtk_init, so by ignoring these options, we simply
confuse the user. This results in transcripts like:

rawtherapee-cli --version
<version>

rawtherapee-cli -v
<full help text>

rawtherapee-cli --help
<version>

rawtherapee-cli -h
<full help text>

(The fact that -v is recognized by the GUI but not the CLI is a bit weird,
but I think it's fine since the CLI unconditionally outputs its version.
But I would like to use -v, -vv, etc., to control verbosity. Currently
there is no verbosity setting.)