main-cli: introduce CliArgs class, currently with one field
Over the next couple of commits I am going to move all the booleans
and other variables that are local to processLineParams into a CliArgs
structure. Then I will move the parsing of CLI arguments into its own
function which constructs this structure and does nothing else. Then
processLineParams will act on that structure.

Because the resulting diff would be so enormous, I am going to do it
piecemeal in small commits where hopefully reviewers can verify that
I'm not changing any logic, even though I'm (necessarily) changing
the names of varibles as they're moved from locals to struct members.

We already sort-of have this, where the 600-line processLineParams
function first processes the CLI arguments and then constructs a job.
But (a) it would be better if we didn't have a mega-function that
does so many things; and (b) the separation is badly incomplete since
several arguments affect the way that other arguments are interpreted
(which is why we insist that -c always be the last argument).

This parsing/processing separation will clean up the code and also let
us fix several longstanding usability issues with the CLI:

* The "streaming" parser lets the user use multiple instances of -j,
  -t, -n, etc., which clobber each other and only the last one actually
  takes effect.
* Similar with -o and -O
* The -c argument must occur last; this is documented but if the user
  fails to do it, the result is not an error but instead surprising
  or wrong behavior.
* The logic for when we show the help and/or do nothing appears and/or
  show a meaningful error message, feel random and arbitrary to the user.