main-cli.cc: clean up "eclipse mode"
In the CLI command-line we have a compile-time flag ECLIPSE_ARGS which
strips the first and last character from all command-line arguments.
This was introduced in fb5466bc8c51c0d4fafdf3a39d3b33310bb4e5a6 which
was part of the rt-cli2 branch and whose provenance I don't really
know.

Probably this flag should just be removed, since you can't actually
set it without editing the code (in two places), and it has a comment
that suggests the author was unsure about it and just hacking around
a specific problem he had. But for now, leave it in and abstract the
code so that we don't have #if ECLIPSE_ARGS laying around all over
the place.

While we're at it, add a length check so that if you do compile this
in, and run it from outside of Eclipse, you don't get unhandled
exceptions which look like

terminate called after throwing an instance of 'std::out_of_range'
  what():  basic_string::at: __n (which is 0) >= this->size() (which is 0)
Aborted (core dumped)

on my system. I changed it so that if there aren't enough characters to
strip, it just does nothing, which seems as good as anything. And better
than a coredump.