main-cli: use unique_ptr to manage processingParams
This commit is a bit larger than I'd like it to be. It

* Adds a move constructor to PartialProfile and AutoPartialProfile
  to allow constructing and manipulating the object before putting
  it into a vector.
* Changes the type of processingParams from a vector of raw pointers to
  PartialProfile, to a vector of unique pointers to AutoPartialProfile
  and deletes the deleteProcParams helper.
* Changes the type of rawParams and imgParams from a raw pointer to
  a unique pointer and removes all the delete and delete Instance calls.

All these changes wound up being tied to each other by the C++11 type
system so I put them all into one commit.

The use of unique pointers is pretty verbose right now. When we move to
C++ 14 or 17 we will get make_unique which will significantly improve
things. The next commit removes a bit of this boilerplate by using
std::unique_ptr in the return value of loadDynamicProfile.

I attempted to also eliminate the use of new/delete for InitialImage,
but it quickly became a very large project because the correct abstraction
there is std::shared_ptr. Currently we use a homebrew refcounting solution
and the InitialImage type is used in dozens of places and replacing this
would take me too far afield of my original goal in this PR (making the
CLI sensible). If this is well-received I may attempt it in a followup.

Similarly, I attempted to use std::unique_ptr for Imagefloat, which
would eliminate the `delete resultImage`. This seemed straightforward
enough but touched so many files that I also decided to abandon the
effort for now. (It probably also makes sense to just wait for C++14/17
to get make_unique to make this process less annoying.)