main-cli: use an enum for outputType
Right now we are using a string to represent a type with three possible
values. Change this to an enum so that if additional output types are
ever added, the compiler will warn us to update every switch.

The existing code has an unreachable path where we call saveToFile if
outputType ever becomes a value different from the three allowed values.
But this value is always explicitly assigned and cannot ever have such
a value.