main-cli: don't set global options
You can see by using `git grep` that none of these options are ever used
except in the batch processing queue, which is not used by the CLI. (The
CLI just calls processImage directory on jobs that it runs through in a
loop.) So these assignments have no effect.

If you don't buy (or don't want to verify) the above argument, there's
an easier way to see that these are useless: they are set only when the
code encounters a CLI argument that doesn't start with - and isn't
consumed by an argument that starts with -. In this case, we assign
the argv1 variable, which is never read EXCEPT to abort the program
with return value 1 if it's nonempty.

So even if these options -did- have an effect, the user would only be
able to trigger them by providing an extra empty-string argument to the
CLI. I don't believe that anybody is doing this on purpose :).